requirements.txt
streamlit>=1.33.0
pypdf>=4.2.0
reportlab>=4.1.0
Pillow>=10.0.0
